+++ packages/components/molecules/f-rating/stories/Rating.stories.js
@@ -22,7 +22,7 @@ export const RatingComponent = (args, { argTypes }) => ({
                     v-bind="$props"
                     :starRatingSize="rating.starRatingSize"
                     :starRating="rating.starRating"
-                    :totalReviews="rating.totalReviews"/>
+                    :reviewCount="rating.reviewCount"/>
             </div>
         </div>`
 });
@@ -51,20 +51,20 @@ RatingComponent.argTypes = {
 RatingComponent.args = {
     ratingVariants: [
         [
-            { starRating: 0, starRatingSize: 'small', totalReviews: 499 },
-            { starRating: 2.3, starRatingSize: 'small', totalReviews: 499 },
-            { starRating: 4.5, starRatingSize: 'small', totalReviews: 499 },
-            { starRating: 1.3, starRatingSize: 'small', totalReviews: 499 },
+            { starRating: 0, starRatingSize: 'small', reviewCount: 499 },
+            { starRating: 2.3, starRatingSize: 'small', reviewCount: 499 },
+            { starRating: 4.5, starRatingSize: 'small', reviewCount: 499 },
+            { starRating: 1.3, starRatingSize: 'small', reviewCount: 499 },
 
-            { starRating: 0, starRatingSize: 'medium', totalReviews: 499 },
-            { starRating: 2.3, starRatingSize: 'medium', totalReviews: 499 },
-            { starRating: 4.5, starRatingSize: 'medium', totalReviews: 499 },
-            { starRating: 1.3, starRatingSize: 'medium', totalReviews: 499 },
+            { starRating: 0, starRatingSize: 'medium', reviewCount: 499 },
+            { starRating: 2.3, starRatingSize: 'medium', reviewCount: 499 },
+            { starRating: 4.5, starRatingSize: 'medium', reviewCount: 499 },
+            { starRating: 1.3, starRatingSize: 'medium', reviewCount: 499 },
 
-            { starRating: 0, starRatingSize: 'large', totalReviews: 499 },
-            { starRating: 2.3, starRatingSize: 'large', totalReviews: 499 },
-            { starRating: 4.5, starRatingSize: 'large', totalReviews: 499 },
-            { starRating: 1.3, starRatingSize: 'large', totalReviews: 499 }
+            { starRating: 0, starRatingSize: 'large', reviewCount: 499 },
+            { starRating: 2.3, starRatingSize: 'large', reviewCount: 499 },
+            { starRating: 4.5, starRatingSize: 'large', reviewCount: 499 },
+            { starRating: 1.3, starRatingSize: 'large', reviewCount: 499 }
         ]
     ]
 };